[feat3]:Comments resolved
