+++ src/common-components/DraggbleTabs.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { useSortable } from "@dnd-kit/sortable";
+import { CSS } from "@dnd-kit/utilities";
+
+export const DraggableTabNode = ({ ...props }) => {
+  const { attributes, listeners, setNodeRef, transform, transition } = useSortable({
+    id: props["data-node-key"],
+  });
+  const style = {
+    ...props.style,
+    transform: CSS.Translate.toString(transform),
+    transition,
+    cursor: "move",
+  };
+  return React.cloneElement(props.children, {
+    ref: setNodeRef,
+    style,
+    ...attributes,
+    ...listeners,
+  });
+};
