+++ src/components/Builder/Education/index.jsx
@@ -4,7 +4,7 @@ import { Button, Col, Form, Input, Row, Space, Tabs } from "antd";
 import { DndContext, PointerSensor, useSensor } from "@dnd-kit/core";
 import { arrayMove, horizontalListSortingStrategy, SortableContext } from "@dnd-kit/sortable";
 import { useGetEducationsQuery } from "../../../api/educationApi";
-import { DraggableTabNode } from "../../../Constants";
+import { DraggableTabNode } from "../../../common-components/DraggbleTabs";
 import { post } from "../../../services/axios";
 import { ResumeContext } from "../../../utils/ResumeContext";
